New Mapping/Waypoint System
Lots of changes in this one:
1)Revamped the mapping system. It now uses a bi-dimensional int array to
store what type each node is. This array more accurately represents a
grid and allows for checking the types of neighboring nodes(useful for
saving with the map editor.
2)Changed the content system. Save content to the Content folder, then
run ContentMover.exe. Select source folder(Content under ContentBuilder
Debug) and destination folder(Content folder under KnightBear debug) and
run to copy over .xnb files. Be sure to build Knightbear before copying.
3)With the new mapping system, Nightmares now have a waypoint system.
Each waypoint is the next node on the path. The code is designed to be
able to read in multiple paths and assign waypoints to new nightmares
according to which path they should take. Currently, all nightmares
default to path 1(index 0).
4)Changes haven't been made to enable the map editor to use the new
mapping system. This is the next task.